ci: adding new Job for Xray in test-ui-e2e workflow + renaming customJsonReporter
