+++ .github/workflows/test-ui-e2e-only-desktop.yml
@@ -38,7 +38,7 @@ on:
         type: boolean
         default: false
       test_execution:
-        description: "Test execution be used in Xray"
+        description: "Test Execution ticket ID"
         required: false
         type: string
         default: "B2CQA-2461"
@@ -63,8 +63,6 @@ jobs:
       CI_OS: "ubuntu-latest"
       PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD: 1
       SPECULOS_IMAGE_TAG: ghcr.io/ledgerhq/speculos:0.9.5
-      XRAY_CLIENT_ID: ${{ secrets.XRAY_CLIENT_ID }}
-      XRAY_CLIENT_SECRET: ${{ secrets.XRAY_CLIENT_SECRET }}
       # DEBUG: "pw:browser*"
       # DEBUG_LOGS: 1
     runs-on: [ledger-live-4xlarge]
@@ -156,9 +154,6 @@ jobs:
   report-and-notify:
     name: "Report and Notify"
     runs-on: [ledger-live-medium]
-    env:
-      XRAY_CLIENT_ID: ${{ secrets.XRAY_CLIENT_ID }}
-      XRAY_CLIENT_SECRET: ${{ secrets.XRAY_CLIENT_SECRET }}
     needs: e2e-tests-linux
     if: always()
     steps:
@@ -190,50 +185,6 @@ jobs:
         if: ${{ !cancelled() }}
         shell: bash
         run: echo "::notice title=Allure report URL::${{ steps.allure-server.outputs.report-url }}"
-
-      - name: Download Xray Reports
-        if: ${{ !cancelled() && inputs.export_to_xray }}
-        uses: actions/download-artifact@v4.1.1
-        with:
-          path: "apps/ledger-live-desktop/artifacts/xray"
-          pattern: xray-reports-*
-          merge-multiple: false
-
-      - name: Extract and Aggregate Xray reports
-        if: ${{ !cancelled() && inputs.export_to_xray }}
-        run: 
-          ./aggregate-xray-reports.sh
-      
-      - name: Upload aggregated xray reports
-        if: ${{ !cancelled() && inputs.export_to_xray }}
-        uses: actions/upload-artifact@v4.3.0
-        with:
-          name: aggregated-xray-reports
-          path: "apps/ledger-live-desktop/artifacts/xray/aggregated-xray-reports.json"
-
-      - name: Authenticate to Xray
-        if: ${{ !cancelled() && inputs.export_to_xray }}
-        id: authenticate
-        run: |
-          response=$(curl -H "Content-Type: application/json" -X POST --data '{"client_id": "${{ secrets.XRAY_CLIENT_ID }}", "client_secret": "${{ secrets.XRAY_CLIENT_SECRET }}"}' https://xray.cloud.getxray.app/api/v2/authenticate)
-          echo "xray_token=$response" >> $GITHUB_OUTPUT
-
-      - name: Publish merged report on Xray
-        if: ${{ !cancelled() && inputs.export_to_xray }}
-        id: publish-xray
-        run: |
-          response=$(curl -H "Content-Type: application/json" \
-                    -H "Authorization: Bearer ${{ steps.authenticate.outputs.xray_token }}" \
-                    -X POST \
-                    --data @apps/ledger-live-desktop/artifacts/xray/aggregated-xray-reports.json \
-                    https://xray.cloud.getxray.app/api/v2/import/execution)
-            key=$(echo $response | jq -r '.key')
-            echo "xray_key=$key" >> $GITHUB_OUTPUT
-
-      - name: Write Xray report in summary
-        if: ${{ !cancelled() && inputs.export_to_xray }}
-        shell: bash
-        run: echo "::notice title=Xray report URL::https://ledgerhq.atlassian.net/browse/${{ steps.publish-xray.outputs.xray_key }}"
 
       - name: Get summary
         if: ${{ !cancelled() }}
@@ -413,3 +364,62 @@ jobs:
           payload-file-path: ${{ github.workspace }}/payload-slack-content.json
         env:
           SLACK_BOT_TOKEN: ${{ secrets.SLACK_LIVE_CI_BOT_TOKEN }}
+
+  upload-to-xray:
+    name: "Upload to Xray"
+    runs-on: [ledger-live-medium]
+    env:
+      XRAY_CLIENT_ID: ${{ secrets.XRAY_CLIENT_ID }}
+      XRAY_CLIENT_SECRET: ${{ secrets.XRAY_CLIENT_SECRET }}
+      XRAY_API_URL: https://xray.cloud.getxray.app/api/v2
+      JIRA_URL: https://ledgerhq.atlassian.net/browse
+    needs: e2e-tests-linux
+    if: ${{ !cancelled() && inputs.export_to_xray }}
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          ref: ${{ inputs.ref || github.sha }}
+      
+      - name: Download Xray Results
+        if: ${{ !cancelled() && inputs.export_to_xray }}
+        uses: actions/download-artifact@v4.1.1
+        with:
+          path: "apps/ledger-live-desktop/artifacts/xray"
+          pattern: xray-reports-*
+          merge-multiple: false
+
+      - name: Extract and Aggregate Xray results
+        if: ${{ !cancelled() && inputs.export_to_xray }}
+        run: 
+          ./aggregate-xray-reports.sh
+      
+      - name: Upload aggregated xray results
+        if: ${{ !cancelled() && inputs.export_to_xray }}
+        uses: actions/upload-artifact@v4.3.0
+        with:
+          name: aggregated-xray-reports
+          path: "apps/ledger-live-desktop/artifacts/xray/aggregated-xray-reports.json"
+
+      - name: Authenticate to Xray
+        if: ${{ !cancelled() && inputs.export_to_xray }}
+        id: authenticate
+        run: |
+          response=$(curl -H "Content-Type: application/json" -X POST --data '{"client_id": "${{ secrets.XRAY_CLIENT_ID }}", "client_secret": "${{ secrets.XRAY_CLIENT_SECRET }}"}' ${{ env.XRAY_API_URL }}/authenticate)
+          echo "xray_token=$response" >> $GITHUB_OUTPUT
+
+      - name: Publish merged report on Xray
+        if: ${{ !cancelled() && inputs.export_to_xray }}
+        id: publish-xray
+        run: |
+          response=$(curl -H "Content-Type: application/json" \
+                    -H "Authorization: Bearer ${{ steps.authenticate.outputs.xray_token }}" \
+                    -X POST \
+                    --data @apps/ledger-live-desktop/artifacts/xray/aggregated-xray-reports.json \
+                    ${{ env.XRAY_API_URL }}/import/execution)
+            key=$(echo $response | jq -r '.key')
+            echo "xray_key=$key" >> $GITHUB_OUTPUT
+
+      - name: Write Xray report in summary
+        if: ${{ !cancelled() && inputs.export_to_xray }}
+        shell: bash
+        run: echo "::notice title=Xray report URL::${{ env.JIRA_URL }}/${{ steps.publish-xray.outputs.xray_key }}"
