+++ apps/ledger-live-desktop/tests/specs/speculos/delete.account.spec.ts
@@ -2,7 +2,7 @@ import { test } from "../../fixtures/common";
 import { specs } from "../../utils/speculos";
 import { Account } from "../../enum/Account";
 import { addTmsLink } from "tests/utils/allureUtils";
-import { getDescription } from "../../utils/custom-json-reporter";
+import { getDescription } from "../../utils/customJsonReporter";
 
 const accounts: Account[] = [
   Account.BTC_1,
