+++ apps/ledger-live-desktop/tests/specs/speculos/subAccount.spec.ts
@@ -2,7 +2,7 @@ import { test } from "../../fixtures/common";
 import { Token } from "../../enum/Tokens";
 import { specs } from "../../utils/speculos";
 import { addTmsLink } from "tests/utils/allureUtils";
-import { getDescription } from "../../utils/custom-json-reporter";
+import { getDescription } from "../../utils/customJsonReporter";
 
 const tokens: Token[] = [
   Token.ETH_USDT,
